[type: refactor] refactor password logic.
