+++ internal/core/openapi/password_test.go
@@ -15,23 +15,22 @@
  * limitations under the License.
  */
 
-package util
+package openapi
 
 import (
 	"testing"
 )
 
-// EncryptPassword Password generation Policy Test
-func TestEncryptPassword(t *testing.T) {
-	password := "admin"
+// saltPassword Password generation Policy Test
+func TestPassword(t *testing.T) {
+	plain := "admin"
 	salt := "07929137ab07437c933d6992321ef9fd"
 
-	encryptPassword := EncryptPassword(password, salt)
+	password := saltPassword(plain, salt)
 
 	// login
-	password = "admin"
-	encryptPasswordLogin := EncryptPassword(password, salt)
-	if encryptPasswordLogin != encryptPassword {
-		t.Errorf("TestEncryptPassword() Login failed salt = %v, password %v", salt, password)
+	plain = "admin"
+	if saltPassword(plain, salt) != password {
+		t.Errorf("TestPassword() Login failed salt = %v, saltPassword %v", salt, plain)
 	}
 }
